# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes et utilisateurs dans l'application Blossom Buddy",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/plants": {
            "get": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Récupérer les plantes de l'utilisateur",
                "description": "Obtenir la liste de toutes les plantes associées à l'utilisateur connecté avec les informations de ville",
                "operationId": "c36bfff3e78962c05ff7eaad8872549c",
                "responses": {
                    "200": {
                        "description": "Liste des plantes de l'utilisateur récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "common_name": {
                                                "type": "string",
                                                "example": "Rose"
                                            },
                                            "watering_general_benchmark": {
                                                "properties": {
                                                    "value": {
                                                        "type": "string",
                                                        "example": "7"
                                                    },
                                                    "unit": {
                                                        "type": "string",
                                                        "example": "days"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "pivot": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID de la relation plant_user",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "city": {
                                                        "description": "Ville où se trouve la plante",
                                                        "type": "string",
                                                        "example": "Paris"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/plant": {
            "post": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Ajouter une plante à l'utilisateur",
                "description": "Associer une plante existante à l'utilisateur connecté avec une ville spécifique",
                "operationId": "c395a1832b5e5cf530cdd39ab8116544",
                "requestBody": {
                    "description": "Données pour ajouter une plante à l'utilisateur",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plant_name",
                                    "city"
                                ],
                                "properties": {
                                    "plant_name": {
                                        "description": "Nom de la plante (recherche partielle)",
                                        "type": "string",
                                        "example": "Rose"
                                    },
                                    "city": {
                                        "description": "Ville où se trouve la plante",
                                        "type": "string",
                                        "example": "Paris"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plante ajoutée avec succès à l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plant added to user successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plante non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Plant]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "plant_name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The plant name field is required."
                                                    }
                                                },
                                                "city": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The city field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/plant/{id}": {
            "delete": {
                "tags": [
                    "User Plants"
                ],
                "summary": "Supprimer une plante de l'utilisateur",
                "description": "Retirer l'association entre l'utilisateur connecté et une plante spécifique",
                "operationId": "3862d76042959edf72ab8837f3cf3b54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relation plant_user (pivot.id)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante supprimée avec succès de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plant deleted from user successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relation plante-utilisateur non trouvée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Plant not found in user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}